backend/requirements.txt
fastapi
uvicorn
python-multipart
openai
faster-whisper
edge-tts
soundfile
numpy
python-dotenv